Refactor: Extract to WhereWithoutIndexMethods
